requirements.txt
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
altair==6.0.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.0
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorlog==6.10.1
dataclasses-json==0.6.7
Deprecated==1.3.1
dill==0.4.0
dirtyjson==1.0.8
docling==2.66.0
docling-core==2.57.0
docling-ibm-models==3.10.3
docling-parse==4.7.2
et_xmlfile==2.0.0
Faker==39.0.0
filelock==3.20.1
filetype==1.2.0
frozenlist==1.8.0
fsspec==2025.12.0
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.3.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
jsonlines==4.0.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
latex2mathml==3.78.1
llama-index-core==0.12.10.post1
llama-index-embeddings-huggingface==0.5.0
llama-index-llms-huggingface-api==0.3.1
llama-index-llms-ollama==0.5.0
llama-index-node-parser-docling==0.3.0
llama-index-readers-docling==0.3.0
llama-index-readers-file==0.4.3
lxml==6.0.2
markdown-it-py==4.0.0
marko==2.2.1
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
mpire==2.10.2
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
mypy_extensions==1.1.0
narwhals==2.14.0
nest-asyncio==1.6.0
networkx==3.6.1
nltk==3.9.2
numpy==2.2.6
ocrmac==1.0.0; platform_system=="Darwin"
ollama==0.6.1
omegaconf==2.3.0
opencv-python==4.12.0.88
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pillow==11.3.0
pluggy==1.6.0
polyfactory==3.2.0
progress_bar==8
propcache==0.4.1
protobuf==6.33.2
psutil==7.2.0
pyarrow==22.0.0
pyclipper==1.4.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pylatexenc==2.10
pyobjc-core==12.1; platform_system=="Darwin"
pyobjc-framework-Cocoa==12.1; platform_system=="Darwin"
pyobjc-framework-CoreML==12.1; platform_system=="Darwin"
pyobjc-framework-Quartz==12.1; platform_system=="Darwin"
pyobjc-framework-Vision==12.1; platform_system=="Darwin"
pypdf==5.9.0
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.0.1
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.3
rapidocr==3.4.5
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rich==14.2.0
rpds-py==0.30.0
rtree==1.4.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
semchunk==2.2.2
sentence-transformers==5.2.0
setuptools==80.9.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
soupsieve==2.8.1
SQLAlchemy==2.0.45
streamlit==1.52.2
striprtf==0.0.26
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
toml==0.10.2
torch==2.9.1
torchvision==0.24.1
tornado==6.5.4
tqdm==4.67.1
transformers==4.57.3
tree-sitter==0.25.2
tree-sitter-c==0.24.1
tree-sitter-java==0.23.5
tree-sitter-javascript==0.25.0
tree-sitter-python==0.25.0
tree-sitter-typescript==0.23.2
typer==0.19.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
watchdog==6.0.0
wheel==0.45.1
wrapt==2.0.1
xlsxwriter==3.2.9
yarl==1.22.0
